+++ lib/flutter_cache_manager.dart
@@ -1,3 +1,5 @@
 export 'src/cache_manager.dart';
-export 'src/file_fetcher.dart';
-export 'src/file_info.dart';
+export 'src/compat/file_fetcher.dart';
+export 'src/result/result.dart';
+export 'src/web/file_service.dart';
+export 'src/web/web_helper.dart' show HttpExceptionWithStatus;